+++ hypervisor/include/arch/x86/cpu.h
@@ -153,7 +153,7 @@
 
 #ifndef ASSEMBLER
 
-int cpu_find_logical_id(uint32_t lapic_id);
+int cpu_find_logical_id(uint8_t lapic_id);
 
 /**********************************/
 /* EXTERNAL VARIABLES             */
